proj/requirements.txt
# Requirements for Python environment to verify p256
pycryptodome==3.22.0
